backend/pyproject.toml
[tool.poetry]
name = "fastapitutorial"
version = "0.0.1"
description = ""
authors = ["Marshall McDonnell <mcdonnellmt@ornl.gov>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.71.0"
uvicorn = "^0.16.0"
Jinja2 = "^3.0.3"
aiofiles = "^0.8.0"
SQLAlchemy = "^1.4.29"
python-dotenv = "^0.19.2"
aiosqlite = "^0.17.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
